qnode12_app/requirements.txt
django>=3.2,<4.0
django-baton
#Wagtail
wagtail==2.16.2

django-social-share
wagtailgmaps
wagalytics==1.3

requests
celery
#gunicorn
uWSGI==2.0.20
redis
django-dotenv
psycopg2-binary
django-storages

#Django Web App dependencies
django-social-auth
social-auth-app-django
sorl-thumbnail
django-widget-tweaks
django-phone-field
django-phonenumber-field[phonenumberslite]
kombu==5.2.2
django-parler
django-localflavor
django-rosetta
wagtail-localize


boto3
#System HOST
PyOpenSSL
python-dotenv==0.21.0

#Ecommerce
braintree
weasyprint
